Moved Filesystem and HDFS into its own project org.backmeup.plugin.api.impl

+++ org.backmeup.plugin.api/src/main/java/org/backmeup/plugin/api/storage/DataObject.java
@@ -6,7 +6,7 @@ public interface DataObject {
 
 	public byte[] getBytes() throws IOException;
 
-	public int getLength();
+	public long getLength();
 	
 	public String getPath();
